refactor(test): rename mongoSetup.js to testSetup.js
