+++ backend/src/routes/api/__tests__/imageApi.test.js
@@ -7,7 +7,7 @@ import Image from "../../../db/models/image.js";
 import {
   useMongoMemoryServer,
   useExpressServer,
-} from "../../../test/mongoSetup.js";
+} from "../../../test/testSetup.js";
 
 jest.mock("firebase-admin"); // Needed to mock the firebase-admin dependency in firebase-auth.js which is in routes
