+++ backend/src/routes/api/__tests__/sceneApi.test.js
@@ -11,7 +11,7 @@ import scenarioAuth from "../../../middleware/scenarioAuth.js";
 import {
   useMongoMemoryServer,
   useExpressServer,
-} from "../../../test/mongoSetup.js";
+} from "../../../test/testSetup.js";
 
 jest.mock("../../../middleware/firebaseAuth");
 jest.mock("../../../middleware/scenarioAuth");
